backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
Django>=5.2
Pillow>=10.0
opencv-python>=4.8.0
numpy>=1.26.0
scikit-learn>=1.3.2
faiss-cpu>=1.7.4
joblib>=1.3.2
scipy>=1.11.4
werkzeug>=3.0.1
PyJWT>=2.8.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Test dependencies (integration tests use requests against Docker backend)
pytest>=7.0.0
pytest-flask>=1.3.0
requests>=2.28.0
